refactored rope (merge request !300)
Squash merge branch 'feature/ryannxue_refactored_rope' into 'master'
refactored rope

+++ src/rope/rope.cu
@@ -178,10 +178,10 @@ __global__ void apply_rotary_pos_emb_prefill_kernel(
   // KV cache layout: [cache_block_idx][pos_in_block][kv_head][head_dim]
   // [warning] : here we assume the cache block is allocated when the token is the first of this
   // new cache block
-  DType *k_cache_row_start = kcache_ptr + cache_block_idx * kcache_block_offset +
+  DType *k_cache_row_start = kcache_ptr + (int64_t)cache_block_idx * (int64_t)kcache_block_offset +
                              pos_in_block * (kNumKVHeads * kQKHeadDim);
-  DType *v_cache_row_start =
-      vcache_ptr + cache_block_idx * vcache_block_offset + pos_in_block * (kNumKVHeads * kVHeadDim);
+  DType *v_cache_row_start = vcache_ptr + (int64_t)cache_block_idx * (int64_t)vcache_block_offset +
+                             pos_in_block * (kNumKVHeads * kVHeadDim);
 
   DType *row_data = load_buffer[iwarp];
 
@@ -429,9 +429,11 @@ __global__ void apply_rotary_pos_emb_decoding_kernel(
           kv_block_indices_ptr[batch_id * max_num_kv_block_per_batch + block_idx_in_batch];
 
       // KV cache layout: [cache_block_idx][pos_in_block][kv_head][head_dim]
-      DType *k_cache_row_start = kcache_ptr + cache_block_idx * kcache_block_offset +
+      DType *k_cache_row_start = kcache_ptr +
+                                 (int64_t)cache_block_idx * (int64_t)kcache_block_offset +
                                  pos_in_block * (kNumKVHeads * kQKHeadDim);
-      DType *v_cache_row_start = vcache_ptr + cache_block_idx * vcache_block_offset +
+      DType *v_cache_row_start = vcache_ptr +
+                                 (int64_t)cache_block_idx * (int64_t)vcache_block_offset +
                                  pos_in_block * (kNumKVHeads * kVHeadDim);
 
       DType *row_data = load_buffer_qkv[iwarp];
@@ -609,7 +611,8 @@ __global__ void apply_rotary_pos_emb_decoding_kernel(
         // Zero K cache
         {
           constexpr int kNumElemPerRow = kNumKVHeads * kQKHeadDim;
-          DType *k_cache_block_start = kcache_ptr + cache_block_idx * kcache_block_offset;
+          DType *k_cache_block_start =
+              kcache_ptr + (int64_t)cache_block_idx * (int64_t)kcache_block_offset;
           // Start from row 1 (skip row 0 as it's being written by compute warps)
           for (int row = 1; row < kv_block_size_divider.divisor; ++row) {
             DType *k_row_pack_ptr = k_cache_block_start + row * kNumKVHeads * kQKHeadDim;
@@ -623,7 +626,8 @@ __global__ void apply_rotary_pos_emb_decoding_kernel(
         // Zero V cache
         {
           constexpr int kNumElemPerRow = kNumKVHeads * kVHeadDim;
-          DType *v_cache_block_start = vcache_ptr + cache_block_idx * vcache_block_offset;
+          DType *v_cache_block_start =
+              vcache_ptr + (int64_t)cache_block_idx * (int64_t)vcache_block_offset;
           // Start from row 1 (skip row 0 as it's being written by compute warps)
           for (int row = 1; row < kv_block_size_divider.divisor; ++row) {
             DType *v_row_pack_ptr = v_cache_block_start + row * kNumKVHeads * kVHeadDim;
